Block sparse operations (#56)
* Support for setting blocks of a block sparse tensor

* Fix tests for new Dictionary design

* Nicer Block printing

+++ src/blocksparse/blockoffsets.jl
@@ -19,8 +19,7 @@ nzblock(block::Block) = block
 
 # Get the offset if the nth block in the block-offsets
 # list
-offset(bofs::BlockOffsets, n::Int) =
-  offset(bofs[n])
+offset(bofs::BlockOffsets, n) = offset(bofs[n])
 
 nnzblocks(bofs::BlockOffsets) = length(bofs)
 nnzblocks(bs::Blocks) = length(bs)
